+++ packages/qr-engine/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @consulta-dev/qr-engine
 
+## 0.1.10
+
 ## 0.1.9
 
 ## 0.1.8